WIP more pushing functionality into splitgraph.core
