+++ splitgraph/core/_objects/utils.py
